+++ primality_deter/AKS.hpp
@@ -8,9 +8,9 @@
 #include <type_traits>
 
 // cette fonction vérifie que n est un "perfect power", si cette fonction renvoie true n est composé sinon aks peut continuer a s'executer
-bool is_perfect_power(unsigned int n);
+bool is_perfect_power(unsigned long int n);
 
-unsigned int maximum (unsigned int a, unsigned int b);
+unsigned long int maximum (unsigned long int a, unsigned long int b);
 
 template <typename T>
 T modpow(T base, T exp, T modulus){
@@ -24,14 +24,14 @@ T modpow(T base, T exp, T modulus){
         return result;
 }
 
-int find_r(int n);
+unsigned long int find_r(unsigned long int n);
 
-unsigned int GCD(unsigned int a, unsigned int b);
+unsigned long int GCD(unsigned long int a, unsigned long int b);
 
-unsigned int EulerPhi(unsigned int n);
+unsigned long int EulerPhi(unsigned long int n);
 
-bool step_5(unsigned int n, unsigned int r);
+bool step_5(unsigned long int n, unsigned long int r);
 
-bool aks(unsigned int n);
+bool aks(unsigned long int n);
 
 #endif
