+++ primality_proba/millerRabin.cpp
@@ -1,8 +1,8 @@
 #include "millerRabin.hpp"
 
 // Modular exponentiation (x^y) % p
-unsigned int expPower(unsigned int x, unsigned int y, unsigned int p){
-								unsigned int res = 1;
+unsigned long int expPower(unsigned long int x, unsigned long int y, unsigned long int p){
+								unsigned long int res = 1;
 								x = x % p;
 								while (y > 0) {
 																if ((y%2) ==  1) {
@@ -16,9 +16,9 @@ unsigned int expPower(unsigned int x, unsigned int y, unsigned int p){
 
 // Returns false if n is composite or
 // true if n is probably prime.
-bool millerTest(unsigned int d, unsigned int n){
-								unsigned int a = 2 + rand() % (n - 4);
-								unsigned int x = expPower(a, d, n);
+bool millerTest(unsigned long int d, unsigned long int n){
+								unsigned long int a = 2 + rand() % (n - 4);
+								unsigned long int x = expPower(a, d, n);
 								if (x == 1 || x == n-1) {
 																return true;
 								}
@@ -34,10 +34,10 @@ bool millerTest(unsigned int d, unsigned int n){
 
 // Returns false if n is composite or true if n
 // is probably prime
-bool millerRabin(unsigned int n, unsigned int k)
+bool millerRabin(unsigned long int n, unsigned long int k)
 {
 								srand (time(NULL));
-								unsigned int d = n - 1;
+								unsigned long int d = n - 1;
 
 								if (n <= 1 || n == 4) return false;
 								if (n <= 3) return true;
@@ -45,7 +45,7 @@ bool millerRabin(unsigned int n, unsigned int k)
 								while (d % 2 == 0)
 																d /= 2;
 
-								for (unsigned int i = 0; i < k; i++)
+								for (unsigned long int i = 0; i < k; i++)
 																if (millerTest(d, n) == false) {
 																								return false;
 																}
